Removed uneeded components. the sats displays now just used the download button directly. Stats slice made smaller
