--- frontend/src/components/dataAggregation/downloadButtons/DownloadCoverageData.js
@@ -1,28 +1,0 @@
-import { useSelector } from "react-redux";
-import DownloadButton from "../../generic/DownloadButton";
-
-/**
- * @description
- * Component for downloading coverage data.
- *
- * @component
- *
- * @example
- * return (
- *   <DownloadCoverageData />
- * )
- */
-const DownloadCoverageData = () => {
-  // Get coverage data.
-  const coverageData = useSelector((state) => state.stats.coverageData);
-
-  return (
-    <DownloadButton
-      data={coverageData}
-      filename="profilerCoverageData"
-      text="Download Coverage Data"
-    />
-  );
-};
-
-export default DownloadCoverageData;
